requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
matplotlib>=3.3.2
tqdm>=4.50.2
pillow>=8.0.1
kagglehub>=0.2.0